feat: set up project, eslint, prettier, typescript
