+++ .gitignore
@@ -14,3 +14,9 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+# build / generate output
+build
+
+# dotenv environment variables file
+.env
